requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.11.0
filelock==3.12.4
fonttools==4.42.1
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.8.0
urllib3==2.0.5
